move caegories service to service dir
