+++ admin-dev/themes/new-theme/js/pages/product/service/categories-service.ts
